fix(docs): Fix provider line parsing in webflow sync. (#3384)
<!-- Describe the problem and your solution --> 
Also add DRYRUN env var for debugging

<!-- Issue ticket number and link (if applicable) -->

<!-- Testing instructions (skip if just adding/editing providers) -->

+++ scripts/webflow-api-sync.ts
@@ -18,6 +18,11 @@ if (!process.env['WEBFLOW_CMS_API_TOKEN']) {
     throw new Error('Missing WEBFLOW_CMS_API_TOKEN');
 }
 
+let dryRun = false;
+if (process.env['DRYRUN']) {
+    dryRun = true;
+}
+
 const webflow = new WebflowClient({ accessToken: process.env['WEBFLOW_CMS_API_TOKEN'] });
 
 const providersPath = 'packages/providers/providers.yaml';
@@ -30,7 +35,7 @@ const files = await fs.readdir(docsPath);
 // we only need a subset of providers based on how our docs are written
 const neededProviders: Record<string, Provider> = {};
 
-const providerLineRegex = /^provider: ([^\s]+)$/m;
+const providerLineRegex = /^provider: ([^\s]+)\s*$/m;
 
 let hasWarnings = false;
 for (const file of files) {
@@ -141,8 +146,11 @@ for (const [slug, provider] of Object.entries(neededProviders)) {
 
         if (!util.isDeepStrictEqual(previous, update)) {
             try {
-                await webflow.collections.items.updateItem(apiCollectionId, item.id, update);
-                console.log(`Updated ${slug}`);
+                if (!dryRun) {
+                    await webflow.collections.items.updateItem(apiCollectionId, item.id, update);
+                }
+
+                console.log(`Updated ${slug} ${dryRun ? '(dry run)' : ''}`);
                 await setTimeout(rateLimitSleep);
             } catch (err) {
                 console.error(`Failed to update ${slug}`, err);
@@ -153,16 +161,18 @@ for (const [slug, provider] of Object.entries(neededProviders)) {
         try {
             const providerCategories: string[] = provider.categories || [];
 
-            await webflow.collections.items.createItem(apiCollectionId, {
-                fieldData: {
-                    name: provider.display_name,
-                    slug: slug,
-                    documentation: provider.docs,
-                    logo: `https://raw.githubusercontent.com/NangoHQ/nango/refs/heads/master/packages/webapp/public/images/template-logos/${slug}.svg`,
-                    'api-categories': providerCategories.map((category) => categoriesBySlug[category]?.id)
-                }
-            });
-            console.log(`Created ${slug}`);
+            if (!dryRun) {
+                await webflow.collections.items.createItem(apiCollectionId, {
+                    fieldData: {
+                        name: provider.display_name,
+                        slug: slug,
+                        documentation: provider.docs,
+                        logo: `https://raw.githubusercontent.com/NangoHQ/nango/refs/heads/master/packages/webapp/public/images/template-logos/${slug}.svg`,
+                        'api-categories': providerCategories.map((category) => categoriesBySlug[category]?.id)
+                    }
+                });
+            }
+            console.log(`Created ${slug} ${dryRun ? '(dry run)' : ''}`);
             await setTimeout(rateLimitSleep);
         } catch (err) {
             console.error(`Failed to update ${slug}`, err);
@@ -178,8 +188,11 @@ for (const toDelete of needDeletion) {
             throw new Error('Unexpected missing item id');
         }
 
-        await webflow.collections.items.deleteItem(apiCollectionId, apiItemsBySlug[toDelete].id);
-        console.log(`Deleted ${toDelete}`);
+        if (!dryRun) {
+            await webflow.collections.items.deleteItem(apiCollectionId, apiItemsBySlug[toDelete].id);
+        }
+
+        console.log(`Deleted ${toDelete} ${dryRun ? '(dry run)' : ''}`);
         await setTimeout(rateLimitSleep);
     } catch (err) {
         console.error(`Failed to delete ${toDelete}`, err);
@@ -187,7 +200,9 @@ for (const toDelete of needDeletion) {
     }
 }
 
-await webflow.sites.publish(siteId, { customDomains: domainIds });
+if (!dryRun) {
+    await webflow.sites.publish(siteId, { customDomains: domainIds });
+}
 
 if (hasWarnings) {
     process.exit(1);
